+++ packages/uikit-core/src/components/presence-selector/presence-selector.vue
@@ -51,12 +51,12 @@ interface PresenceOption {
 }
 
 const statusIconMap: Record<PresenceSelectorValue, string> = {
-  online: 'presence/empty',
-  offline: 'presence/empty',
-  away: 'presence/clock',
-  busy: 'presence/equals',
-  doNotDisturb: 'presence/minus',
-  custom: 'presence/star',
+  online: 'filled/circle/empty',
+  offline: 'filled/circle/empty',
+  away: 'filled/circle/clock',
+  busy: 'filled/circle/equals',
+  doNotDisturb: 'filled/circle/minus',
+  custom: 'filled/circle/star',
 }
 
 const fixedStatusValues = [